move generation files

+++ source/gen/generate.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
-cd source
-rm generate.class
+cd ..
 javac generate.java
 java generate ./
+if [ -f "generate.class" ]; then
+  rm generate.class
+fi
 echo Done.
